Separate Markdown parsing in HTML to separate extension Move Markdown parsing from the extra extension to a separate extension called md_in_html.
