+++ docs/extensions/index.md
@@ -35,25 +35,26 @@ maintained here and all bug reports should be made to the project. If you
 have a typical install of Python-Markdown, these extensions are already
 available to you using the "Entry Point" name listed in the second column below.
 
-Extension                            | Entry Point    | Dot Notation
------------------------------------- | -------------- | ------------
-[Extra]                              | `extra`        | `markdown.extensions.extra`
-&nbsp; &nbsp; [Abbreviations][]      | `abbr`         | `markdown.extensions.abbr`
-&nbsp; &nbsp; [Attribute Lists][]    | `attr_list`    | `markdown.extensions.attr_list`
-&nbsp; &nbsp; [Definition Lists][]   | `def_list`     | `markdown.extensions.def_list`
-&nbsp; &nbsp; [Fenced Code Blocks][] | `fenced_code`  | `markdown.extensions.fenced_code`
-&nbsp; &nbsp; [Footnotes][]          | `footnotes`    | `markdown.extensions.footnotes`
-&nbsp; &nbsp; [Tables][]             | `tables`       | `markdown.extensions.tables`
-[Admonition][]                       | `admonition`   | `markdown.extensions.admonition`
-[CodeHilite][]                       | `codehilite`   | `markdown.extensions.codehilite`
-[Legacy Attributes][]                | `legacy_attr`  | `markdown.extensions.legacy_attr`
-[Legacy Emphasis][]                  | `legacy_em`    | `markdown.extensions.legacy_em`
-[Meta-Data]                          | `meta`         | `markdown.extensions.meta`
-[New Line to Break]                  | `nl2br`        | `markdown.extensions.nl2br`
-[Sane Lists]                         | `sane_lists`   | `markdown.extensions.sane_lists`
-[SmartyPants]                        | `smarty`       | `markdown.extensions.smarty`
-[Table of Contents]                  | `toc`          | `markdown.extensions.toc`
-[WikiLinks]                          | `wikilinks`    | `markdown.extensions.wikilinks`
+Extension                          | Entry Point    | Dot Notation
+---------------------------------- | -------------- | ------------
+[Extra]                            | `extra`        | `markdown.extensions.extra`
+&nbsp; &nbsp; [Abbreviations]      | `abbr`         | `markdown.extensions.abbr`
+&nbsp; &nbsp; [Attribute Lists]    | `attr_list`    | `markdown.extensions.attr_list`
+&nbsp; &nbsp; [Definition Lists]   | `def_list`     | `markdown.extensions.def_list`
+&nbsp; &nbsp; [Fenced Code Blocks] | `fenced_code`  | `markdown.extensions.fenced_code`
+&nbsp; &nbsp; [Footnotes]          | `footnotes`    | `markdown.extensions.footnotes`
+&nbsp; &nbsp; [Markdown in HTML]   | `md_in_html`   | `markdown.extensions.md_in_html`
+&nbsp; &nbsp; [Tables]             | `tables`       | `markdown.extensions.tables`
+[Admonition]                       | `admonition`   | `markdown.extensions.admonition`
+[CodeHilite]                       | `codehilite`   | `markdown.extensions.codehilite`
+[Legacy Attributes]                | `legacy_attr`  | `markdown.extensions.legacy_attr`
+[Legacy Emphasis]                  | `legacy_em`    | `markdown.extensions.legacy_em`
+[Meta-Data]                        | `meta`         | `markdown.extensions.meta`
+[New Line to Break]                | `nl2br`        | `markdown.extensions.nl2br`
+[Sane Lists]                       | `sane_lists`   | `markdown.extensions.sane_lists`
+[SmartyPants]                      | `smarty`       | `markdown.extensions.smarty`
+[Table of Contents]                | `toc`          | `markdown.extensions.toc`
+[WikiLinks]                        | `wikilinks`    | `markdown.extensions.wikilinks`
 
 [Extra]: extra.md
 [Abbreviations]: abbreviations.md
@@ -68,6 +69,7 @@ Extension                            | Entry Point    | Dot Notation
 [Legacy Emphasis]: legacy_em.md
 [Meta-Data]: meta_data.md
 [New Line to Break]: nl2br.md
+[Markdown in HTML]: md_in_html.md
 [Sane Lists]: sane_lists.md
 [SmartyPants]: smarty.md
 [Table of Contents]: toc.md
